+++ src/main/java/cn/nukkit/item/ItemPotionSplash.java
@@ -40,7 +40,7 @@ public class ItemPotionSplash extends ProjectileItem {
 
     @Override
     protected void correctNBT(CompoundTag nbt) {
-        nbt.putInt("PotionId", this.getDamage());
+        nbt.putShort("PotionId", this.getDamage());
 
         nbt.putCompound("Item", NBTIO.putItemHelper(this));
     }
